Rename spi_app and get building
