+++ esp/mqtt5/main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "app_main.c" "spi_app.c"
+idf_component_register(SRCS "app_main.c" "fmt_esp_spi.c"
                     INCLUDE_DIRS "." "../firmware")
